Refactor: Apply consistent naming conventions to src files for better code readability
