+++ src/schema/v1/bid/bid.apply.ts
