+++ src/schema/v1/bid/bid.delete.ts
