+++ src/schema/v1/consultant/consultant.delete.ts
