+++ src/schema/v1/domain/domain.getAll.ts
