+++ src/schema/v1/faq/faq.delete.ts
