+++ src/schema/v1/freelancer/freelancer.delete.ts
