+++ src/schema/v1/hireDehixTalent/hireDehixTalent.delete.ts
