+++ src/schema/v1/project/project.delete.ts
