+++ src/schema/v1/skills/skills.getAll.ts
