+++ src/types/v1/faq/deleteFaq.ts
