+++ src/types/v1/freelancer/deleteFreelancer.ts
